player: split backend modules
